src/org/launchcode/techjobs_oo/tests/JobTest.java
package org.launchcode.techjobs_oo.tests;

import org.junit.Before;
import org.junit.Test;
import org.launchcode.techjobs_oo.*;

import static org.junit.Assert.*;

public class JobTest {
    Job emptyJob1;
    Job emptyJob2;
    Job completeJob1;
    Job completeJob2;

    @Before
    public void createJobObjects() {
        emptyJob1 = new Job();
        emptyJob2 = new Job();
        completeJob1 = new Job("Product tester", new Employer("ACME"), new Location("Desert"), new PositionType("Quality control"), new CoreCompetency("Persistence"));
        completeJob2 = new Job("Product tester", new Employer("ACME"), new Location("Desert"), new PositionType("Quality control"), new CoreCompetency("Persistence"));

    }

    @Test
    public void testSettingJobId() {
        assertEquals(emptyJob2.getId()-1, emptyJob1.getId(), 0);
    }

    @Test
    public void testJobConstructorSetsAllFields() {
        assertEquals("Product tester", completeJob1.getName());
        assertEquals("ACME", completeJob1.getEmployer().toString());
        assertTrue(completeJob1.getEmployer() instanceof Employer);
        assertEquals("Desert", completeJob1.getLocation().toString());
        assertTrue(completeJob1.getLocation() instanceof Location);
        assertEquals("Quality control", completeJob1.getPositionType().toString());
        assertTrue(completeJob1.getPositionType() instanceof PositionType);
        assertEquals("Persistence", completeJob1.getCoreCompetency().toString());
        assertTrue(completeJob1.getCoreCompetency() instanceof CoreCompetency);
    }

    @Test
    public void testJobsForEquality() {
        assertFalse(completeJob1.equals(completeJob2));
    }

    @Test
    public void testJobToStringAllFields() {
        assertEquals("\nID: 3\n" +
                "Name: Product tester\n" +
                "Employer: ACME\n" +
                "Location: Desert\n" +
                "Position Type: Quality control\n" +
                "Core Competency: Persistence\n", completeJob1.toString());
    }

    @Test
    public void testJobToStringMissingFields() {
        Job missingFieldsJob = new Job("", new Employer("ACME"), new Location("Desert"), new PositionType("Quality control"), new CoreCompetency("Persistence"));
        assertEquals("\nID: 13\n" +
                "Name: Data not available\n" +
                "Employer: ACME\n" +
                "Location: Desert\n" +
                "Position Type: Quality control\n" +
                "Core Competency: Persistence\n", missingFieldsJob.toString());
    }

// TODO: fix nullpointerexception in Job.java toString() method to pass this test

//    @Test
//    public void testJobToStringOnlyIdField() {
//        assertEquals(emptyJob1.toString(), "OOPS! This job does not seem to exist.");
//    }

}
